requirements.txt
3to2==1.1.1
alembic==0.9.9
asn1crypto==0.24.0
Babel==2.5.1
blinker==1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
decorator==4.1.2
dominate==2.3.1
dotenv==0.0.5
enum34==1.1.6
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SocketIO==2.9.6
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
gevent==1.3.6
greenlet==0.4.15
guess-language-spirit==0.5.3
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pbr==3.1.1
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
python-engineio==2.0.3
python-socketio==1.9.0
pytz==2017.3
requests==2.19.1
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.23
visitor==0.1.3
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
